+++ migrations/20250831230214_glassdoor_add_column.ts
@@ -1,16 +1,23 @@
 import type { Knex } from "knex";
 
-
 export async function up(knex: Knex): Promise<void> {
-	await knex.schema.table('glassdoor', table => {
-		table.string('region');
-	});
-}
+  const hasColumn = await knex.schema.hasColumn('glassdoor', 'region');
+  if (hasColumn) {
+    await knex.schema.alterTable('glassdoor', table => {
+      table.dropColumn('region');
+    });
+  }
 
+  await knex.schema.alterTable('glassdoor', table => {
+    table.string('region');
+  });
+}
 
 export async function down(knex: Knex): Promise<void> {
-	await knex.schema.table('glassdoor', table => {
-		table.dropColumn('region');
-	});
+  const hasColumn = await knex.schema.hasColumn('glassdoor', 'region');
+  if (hasColumn) {
+    await knex.schema.alterTable('glassdoor', table => {
+      table.dropColumn('region');
+    });
+  }
 }
-
